knowledge-graph-api/requirements.txt
# HTTP client per Ollama
httpx>=0.27.0

# Neo4j
neo4j>=5.18.0

# Redis for AI
redis[hiredis]>=5.0.0

# FastAPI
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# Pydantic
pydantic>=2.7.0
pydantic-settings>=2.3.0

# Document processing
pypdf>=4.2.0
python-docx>=1.1.0

# Utilities
tenacity>=8.3.0
structlog>=24.1.0
python-dotenv>=1.0.0
numpy>=1.26.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-mock>=3.14.0

# Linting
ruff>=0.4.0
